requirements.txt
Pillow
rembg
onnxruntime